rename to k8s-client
