+++ Dockerfile
@@ -3,7 +3,7 @@ FROM ruby:2.4.3
 WORKDIR /app
 
 COPY Gemfile *.gemspec ./
-COPY lib/pharos/kube/client/version.rb ./lib/pharos/kube/client/
+COPY lib/k8s/client/version.rb ./lib/k8s/client/
 RUN bundle install
 
 COPY . .
