+++ lib/k8s-client.rb
@@ -1,0 +1,13 @@
+require 'k8s/api/metav1'
+require 'k8s/api/version'
+
+require 'k8s/config'
+require 'k8s/logging'
+
+require 'k8s/api_client'
+require "k8s/client"
+require "k8s/error"
+require 'k8s/resource'
+require 'k8s/resource_client'
+require 'k8s/stack'
+require 'k8s/transport'
