+++ lib/k8s/client/version.rb
@@ -1,0 +1,5 @@
+module K8s
+  class Client
+    VERSION = "0.1.0"
+  end
+end
