--- lib/pharos-kube-client.rb
@@ -1,13 +1,0 @@
-require 'pharos/kube/api/metav1'
-require 'pharos/kube/api/version'
-
-require 'pharos/kube/config'
-require 'pharos/kube/logging'
-
-require 'pharos/kube/api_client'
-require "pharos/kube/client"
-require "pharos/kube/error"
-require 'pharos/kube/resource'
-require 'pharos/kube/resource_client'
-require 'pharos/kube/stack'
-require 'pharos/kube/transport'
