--- lib/pharos/kube/api.rb
@@ -1,28 +1,0 @@
-require 'dry-types'
-require 'dry-struct'
-
-module Pharos
-  module Kube
-    module API
-      module Types
-        include Dry::Types.module
-      end
-
-      class Struct < Dry::Struct
-        # input from JSON with string keys
-        transform_keys(&:to_sym)
-
-        # @param data [Hash]
-        # @return [self]
-        def self.from_json(data)
-          return new(data)
-        end
-
-        # @return [String]
-        def to_json
-          to_hash.to_json
-        end
-      end
-    end
-  end
-end
