--- lib/pharos/kube/client/version.rb
@@ -1,7 +1,0 @@
-module Pharos
-  module Kube
-    class Client
-      VERSION = "0.1.0"
-    end
-  end
-end
