+++ pharos-kube-client.gemspec
@@ -1,16 +1,16 @@
 
 lib = File.expand_path("../lib", __FILE__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require "pharos/kube/client/version"
+require "k8s/client/version"
 
 Gem::Specification.new do |spec|
-  spec.name          = "pharos-kube-client"
-  spec.version       = Pharos::Kube::Client::VERSION
+  spec.name          = "k8s-client"
+  spec.version       = K8s::Client::VERSION
   spec.authors       = ["Kontena, Inc."]
   spec.email         = ["info@kontena.io"]
 
   spec.summary       = "Kubernetes client library"
-  spec.homepage      = "https://github.com/kontena/pharos-kube-client"
+  spec.homepage      = "https://github.com/kontena/k8s-client"
 
   # Specify which files should be added to the gem when it is released.
   # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
