+++ spec/pharos/kube/stack_spec.rb
@@ -1,8 +1,8 @@
-RSpec.describe Pharos::Kube::Stack do
+RSpec.describe K8s::Stack do
   include FixtureHelpers
 
   context "for a simple whoami stack" do
-    let(:client) { instance_double(Pharos::Kube::Client) }
+    let(:client) { instance_double(K8s::Client) }
 
     subject { described_class.load(fixture_path('stacks/whoami.yaml'), name: 'whoami') }
